Undo of moved elements bug fixed

+++ ACS/ACS/MainWindow.xaml.cs
@@ -7653,8 +7653,8 @@ namespace Asterics.ACS {
             CommandObject co = new CommandObject("moveComponent", selectedComponentList.ToArray());
             for (int i = 0; i < selectedComponentList.Count; i++) {
                 componentType mc = selectedComponentList.ElementAt(i);
-                co.Parameter.Add(double.Parse(mc.layout.posX));
-                co.Parameter.Add(double.Parse(mc.layout.posY));
+                co.Parameter.Add(int.Parse(mc.layout.posX));
+                co.Parameter.Add(int.Parse(mc.layout.posY));
             }
             return co;
         }
